Add rendering for item list
This amends the homepage route to pass listing data to the handlebars, which now loops through the input data and renders them in the carousel. The image locations have also been amended in order to be rendered in child directories and styling has been applied to tidy up the homepage display of items. This all results in a logged out user being able to see a summary of listed items.